+++ r5a/src/main/java/alma/logoot/network/NetworkService.java
@@ -7,5 +7,5 @@ import com.google.gwt.user.client.rpc.RemoteServiceRelativePath;
 @RemoteServiceRelativePath("logoot")
 public interface NetworkService extends RemoteService{
 	void send(String o);
-	int register();	
+	long register();	
 }
